Refactor answerQuiz with helper function
Add checkTypeAnswer to answerQuiz
