+++ src/helpers/quizActionHelpers.js
@@ -1,4 +1,11 @@
-import store from '../store';
+import store from "../store";
+
+import removeDiacritics from "../helpers/removeDiacritics";
+
+const simple = str =>
+  removeDiacritics(str.toLowerCase())
+    .replace(/\u002D/g, " ")
+    .replace(/[^\w\s]/g, "");
 
 export const generateAnswerArray = filterRegions => {
   const quizAnswers = [...filterRegions];
@@ -15,15 +22,15 @@ export const generateQuizState = (quizAnswers, quizType) => ({
   isQuizActive: true,
   activeQuestionNum: 0,
   quizGuesses: [],
-  selectedProperties: '',
-  isTypeQuizActive: quizType.split('_')[0] === 'type',
+  selectedProperties: "",
+  isTypeQuizActive: quizType.split("_")[0] === "type"
 });
 
 export const checkClickAnswer = ansGeoProperties => {
   const {
     activeQuestionNum,
     quizAnswers,
-    selectedProperties,
+    selectedProperties
   } = store.getState().quiz;
   const { regionKey } = store.getState().map;
   const isAnswerCorrect =
@@ -31,6 +38,27 @@ export const checkClickAnswer = ansGeoProperties => {
   const newGeoProperties = isAnswerCorrect
     ? ansGeoProperties
     : selectedProperties;
+  return { isAnswerCorrect, newGeoProperties };
+};
+
+export const checkTypeAnswer = userGuess => {
+  const { quizAnswers, activeQuestionNum, quizType } = store.getState().quiz;
+  const { geographyPaths } = store.getState().data;
+  const { regionKey } = store.getState().map;
+
+  const answerProperties = geographyPaths.find(
+    geo => geo.properties[regionKey] === quizAnswers[activeQuestionNum]
+  ).properties;
+
+  const isAnswerCorrect =
+    quizType.split("_")[1] === "name"
+      ? answerProperties.spellings.some(
+          name => simple(userGuess) === simple(name)
+        )
+      : simple(userGuess) === simple(answerProperties.capital);
+
+  const newGeoProperties = isAnswerCorrect ? answerProperties : "";
+
   return { isAnswerCorrect, newGeoProperties };
 };
